Make player health a float

+++ Assets/Scripts/Enemy.cs
@@ -7,7 +7,7 @@ public class Enemy : MonoBehaviour {
     
     #region Inspector properties
     public float MoveSpeed;
-    public int   AttackDamage;
+    public float AttackDamage;
     public float DamageApplicationInterval;
     #endregion
