+++ Assets/Scripts/GameMaster.cs
@@ -50,7 +50,7 @@ public class GameMaster : MonoBehaviour {
         }
     }
 
-    private void OnPlayerHealthChanged(int newHealth, Player.HealthChangeCause cause) {
+    private void OnPlayerHealthChanged(float newHealth, Player.HealthChangeCause cause) {
         if (newHealth <= 0) {
             Debug.Log("Player Dead");
             SceneManager.LoadScene(SceneManager.GetActiveScene().name);
